correction apres premiere correction 19

+++ srcs/convert/convert_unsigned.c
@@ -12,16 +12,16 @@
 
 #include "ft_printf.h"
 
-static int	set_precision_len(int precision, char *tmp)
+static long	set_precision_len(long precision, char *tmp)
 {
-	if (precision <= ft_strclen(tmp, ' '))
+	if (precision <= (long)ft_strclen(tmp, ' '))
 		return (-1);
 	else
 	{
-		if (precision >= ft_strlen(tmp))
+		if (precision >= (long)ft_strlen(tmp))
 			return (precision);
 		else
-			return ((int)ft_strlen(tmp));
+			return ((long)ft_strlen(tmp));
 	}
 	return (0);
 }
@@ -45,7 +45,7 @@ static int	precision(t_data *data)
 	if (length == precision)
 		ft_memcpy(&(data->value_format[precision - ft_strclen(tmp, ' ')]),
 			tmp, ft_strclen(tmp, ' '));
-	else if (length == ft_strlen(tmp))
+	else if (length == (long)ft_strlen(tmp))
 		ft_memcpy(&(data->value_format[precision - ft_strclen(tmp, ' ')]),
 			tmp, (length - precision) + ft_strclen(tmp, ' '));
 	return (1);
